requirements.txt
asgiref==3.7.2
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==2.0.12
colorama==0.4.6
decorator==5.1.1
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
iniconfig==2.0.0
ipython==8.18.1
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pluggy==0.13.1
prompt-toolkit==3.0.43
pure-eval==0.2.2
py==1.11.0
Pygments==2.17.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.13
